Split gmregclear at original constant pools
Recover the Adventure and All-Star percentage accessors by assigning their duplicated conversion constants to separate translation units. Keep fn_8017F008 with regclear so its state remains private and its BSS ordering stays exact.

Define the shared game-speed constants in gm_17C0; volatile reads prevent MWCC from folding duplicate literals while preserving all existing matches.
