+++ configure.py
@@ -1036,6 +1036,9 @@ config.libs = [
             Object(Matching, "melee/gm/gm_17BA.c"),
             Object(Matching, "melee/gm/gmregcommon.c"),
             Object(NonMatching, "melee/gm/gm_17C0.c"),
+            Object(NonMatching, "melee/gm/gm_17C9.c"),
+            Object(NonMatching, "melee/gm/gm_17E4.c"),
+            Object(NonMatching, "melee/gm/gm_17EB.c"),
             Object(NonMatching, "melee/gm/gmregclear.c"),
             Object(NonMatching, "melee/gm/gm_1832.c"),
             Object(NonMatching, "melee/gm/gmtoulib.c"),
